refactoring so that the form is separate component from the modal
